remove legacy reward calculations
