+++ packages/app/src/contexts/Payouts/defaults.ts
@@ -4,8 +4,6 @@
 
 import type { PayoutsContextInterface } from './types'
 
-export const MaxSupportedPayoutEras = 7
-
 export const defaultPayoutsContext: PayoutsContextInterface = {
   unclaimedRewards: {
     total: '0',
